ref: separate build and running logic
